+++ Code/Prosjekt 2/CAN_Pong_v2/Paddle.h
@@ -2,6 +2,7 @@
 #define Paddle_h
 
 #define SCREEN_HEIGHT 64
+#include "Arduino.h"
 
 enum side
 {
@@ -12,10 +13,10 @@ enum side
 class Paddle
 {
 public:
-  int paddleWidth;
-  int paddleHeight;
-  int paddle_x;
-  int paddle_y;
+  uint8_t paddleWidth;
+  uint8_t paddleHeight;
+  uint8_t paddle_x;
+  uint8_t paddle_y;
 
   Paddle(side paddleSide);
   ~Paddle();
